% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_typing.R
\name{applyType}
\alias{applyType}
\title{Type Application}
\usage{
applyType(input_type)
}
\arguments{
\item{input_type}{The type prefix for the triplet type}
}
\value{
the formatter function to formally type the input var
}
\description{
Type Application
}
