% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{makeRootSkillCluster}
\alias{makeRootSkillCluster}
\title{Make RS Cluster}
\usage{
makeRootSkillCluster(skill_base_name, base_difficulty)
}
\arguments{
\item{skill_base_name}{The base name of the skill, will be appended with the rs- prefix}

\item{base_difficulty}{The base difficulty of that type of skill}
}
\description{
Make RS Cluster
}
