% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{makeSpecificSkillCluster}
\alias{makeSpecificSkillCluster}
\title{Make SS Cluster}
\usage{
makeSpecificSkillCluster(skill_spec, skill_base_name, skill_weight,
  skill_desc = NULL, skill_comp_date = NULL)
}
\arguments{
\item{skill_spec}{The name of the specific skill}

\item{skill_base_name}{The name of the base skill}

\item{skill_weight}{The weight of the specific skill}

\item{skill_desc}{The description of the specific skill}

\item{skill_comp_date}{The date of completion for the specific skill}
}
\description{
Make SS Cluster
}
