% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_typing.R
\name{reTypeVect}
\alias{reTypeVect}
\title{Retype Vectorised}
\usage{
reTypeVect(vect)
}
\arguments{
\item{char}{the formally typed character element}
}
\description{
Function to re-apply the internal typeing of the stored var
}
