% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{DS Skill stream test dataset}
\format{A tibble with 15 obs. of 3 variables:
\describe{
  \item{Origin}{The root object, to which a value can be attached}
  \item{Relationship}{The relationship tying the root to the value}
  \item{Value}{The values tied into the object}
  ...
}}
\usage{
test_data
}
\description{
A testing dataset which gives a triplet database for examples use
}
\keyword{datasets}
