% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{verifySpecificSkillCluster}
\alias{verifySpecificSkillCluster}
\title{Verify SS Cluster}
\usage{
verifySpecificSkillCluster(skill_spec, verify_by)
}
\arguments{
\item{skill_spec}{The name of the specific skill}

\item{verify_by}{The Verifying party}
}
\description{
Verify SS Cluster
}
